+++ src/main/java/dongguk/osori/domain/follow/FollowRepository.java
@@ -1,10 +1,9 @@
 package dongguk.osori.domain.follow;
 
+import dongguk.osori.domain.follow.entity.Follow;
 import dongguk.osori.domain.user.entity.User;
-import io.lettuce.core.dynamic.annotation.Param;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
